fix: corrigido problema em conciliacao ao upar balancete
fix: alterado para long id

+++ src/main/java/com/clearcont/clearcontapp/repository/BalanceteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BalanceteRepository extends JpaRepository<Balancete, Integer> {
+public interface BalanceteRepository extends JpaRepository<Balancete, Long> {
     List<Balancete> findBalanceteByEmpresa_IdAndMesAndAno(Integer id, String month, Integer year);
     void deleteAllByEmpresa_Id(Integer id);
 }
