+++ src/main/java/com/clearcont/clearcontapp/service/BalanceteService.java
@@ -10,35 +10,35 @@ import java.util.List;
 @Service
 public class BalanceteService {
     private final BalanceteRepository repository;
-    
+
     public BalanceteService(BalanceteRepository repository) {
         this.repository = repository;
     }
-    
+
     public List<Balancete> getAll() {
         return repository.findAll();
     }
-    
-    public Balancete getById(Integer balanceteId) {
+
+    public Balancete getById(Long balanceteId) {
         return repository.findById(balanceteId).get();
     }
-    
+
     public List<Balancete> getByCompanyAndPeriod(Integer id, String mes, Integer ano) {
         return repository.findBalanceteByEmpresa_IdAndMesAndAno(id, mes, ano);
     }
-    
+
     public void delete(Balancete balancete) {
         repository.delete(balancete);
     }
-    
+
     public Balancete update(Balancete balancete) {
         return repository.saveAndFlush(balancete);
     }
-    
+
     public Balancete save(Balancete balancete) {
         return repository.save(balancete);
     }
-    
+
     @Transactional
     public void saveAll(Integer id, List<Balancete> balancetes) {
         repository.deleteAllByEmpresa_Id(id);
